Bug x - Move the APZCTreeManager initialization for the GPU process to CompositorBridgeParent initialization. r?
This is important because the RecvInitialize method is run via a sync
IPC message, and so we know that when return to the caller in the UI
process, that the APZCTreeManager will have been created. This ensures
that when we create the APZInputBridge actors (which happens on a
different top-level protocol, and so can run concurrently) we know that
the concrete APZCTreeManager is ready for use.