Bug 1437036 - Stop the reftest harness from waiting for MozAfterPaint during infinite/superlong animations. r?hiro
These tests rely on an optimization within Gecko where it stops firing
MozAfterPaint events if there was no visible change to the generated
layers. This allows the reftest harness to exit the
waiting-for-MozAfterPaint loop and proceed with the test. However, with
webrender, this optimization does not exist and so the loop never exits.
In order to solve this problem, this patch adds an explicit mechanism to
exit the loop by means of a class attribute on the root element of the
test page.

MozReview-Commit-ID: 17ta5kLPDr9