Bug x - Stop using SamplingFilter for WebRender. r?jrmuizel
SamplingFilter has no way to represent the "crisp edges" value so using
it results in a loss of information that we need. This change is a
refactoring that adds a new nsLayoutUtils function that effectively
combines the existing function to get the sampling filter with the
function to convert the sampling filter to an ImageRendering. No
functional changes in this patch.