Bug 1448513 - Only do the panThreshold check if the threshold is > 0. r?botond
This optimization also sneakily avoids the call to UpdateWithTouchAtDevicePoint
which is what updates the Axis::mPos variable to the new touch point. Without
that update, mPos remains at the touchstart point, which ensures that the full
pixel delta from all the touchmove events gets turned into scrolling. Without
this, the first touchmove event always gets consumed in the process of overcoming
the pan threshold, which lead to flaky test behaviour.