Bug 1505375 - Send the interleaved WR display list dump to logcat. r?nical
Using the gfx.webrender.dl.dump-* prefs is supposed to print a display
list dump with the gecko and WR display items appropriately interleaved.
However on Android the gecko items get sent to logcat via printf_stderr,
and the WR display items get sent to stderr, and so get lost. This makes
the WR items also go to logcat so that they show up properly.

Differential Revision: https://phabricator.services.mozilla.com/D11196