Bug 1638664 - Improve the way MozsearchIndexer merges analysis data. r?asuth,emilio
Instead of doing this:
a) read existing file into memory
b) append new entries
c) sort all entries
d) write unique entries back to file

We now do this:
a) sort new entries
b) loop through existing file entries, one at a time, writing them to a tmp file
c) insert the new entries in between the existing file entries in lexicographic order,
   and deduplicating the new entries
d) write any remaining new entries (that are lexicographically after the last
   entry in the pre-existing file), again deduplicating the entries
e) move tmp file back to original file location

This avoids reading the entire file into memory which could be potentially
hundreds of MB large.

The changes in FileOperations.* are needed to support these changes, as we now
have two files that we're dealing with - reading from one and writing to the
other. We still use a Mutex (Windows) or exclusive-flock (POSIX) on the file
for the duration of the entire operation, so we should still be robust in the
face of multiple concurrently running clang processes.