Bug x - Make a bunch of tests not fail with apz.allow_zooming set to true by default. r?tnikkel
In some cases where meta-viewport is enabled, we can enable overlay scrollbars
to make the test pass, because the test fails because of the unsupported
combination of layout scrollbars + meta-viewport + APZ zooming.

In other cases we just turn of APZ zooming on the individual test until we
can fix it properly.