Bug 1655733 - Introduce a ScrollPositionUpdate and plumb it in. r=tnikkel!,botond!
This adds a ScrollPositionUpdate class. Code in nsGfxScrollFrame creates
instances of these classes and saves them to an array. The array is sent
in the scroll metadata to the compositor as part of a paint transaction.
Currently this data is not used, and exists alongside (independently of)
the existing scroll fields, so this patch should not have any functional
changes.