Bug 1662013 - Make the scroll-behavior-element WPT test pass. r=tnikkel!
The previous patches caused the scroll-behavior-element WPT to fail, specificallly
the subtest that does a no-op instant scroll immediately following an APZ smooth
scroll, within the same transaction. In this subtest, the APZ smooth scroll
request gets added to mScrollUpdates but then the instant scroll is a no-op
(destination is same as current position) so we early return and don't add a
scroll update for it. This patch adds a bit of code to the early-exit path to
detect this scenario and not take the early exit. By itself this causes a
different test to fail, because we also hit this codepath when the mainthread
internally does a no-op instant scroll to reclamp the scroll offset after a
reflow. That requires further refining the condition with a new scroll origin
used for the clamping call.

This ain't the prettiest thing but I'm hopeful that after untangling more of
the surrounding code in the future we can replace this with something more
elegant.

Differential Revision: https://phabricator.services.mozilla.com/D88745