Bug 1648491 - Add a touch-action root tracker on the nsDisplayListBuilder. r?miko
This patch adds some plumbing so that the nsDisplayListBuilder can track a
"touch-action root" frame, which is defined as the nearest ancestor frame which
modified the touch-action bits in the hit-test info. This field is currently
initialized on constructing the nsDisplayListBuilder, but not yet updated (that
will happen in the next patch). The plumbing also propagates the current value
of the field into any hit-test info items that are created, so that they can
be used by the FrameLayerBuilder when collapsing touch-action regions (also in
a future patch).