Bug 1231177 - Allow scroll position to advance when spamming multiple scrollframes in an interleaved manner. r?botond
In the case where there are two scrollframes that get scrolled in an
interleaved manner, the scroll generation counter continually advances
and blocks APZ from updating the main-thread scroll position. This
fixes that by removing the scroll generation counter check from the
relevant condition. The clause is entirely redundant when there is only
a single scrollframe and only causes problems when there are multiple
scrollframes.

Differential Revision: https://phabricator.services.mozilla.com/D97034